NEW dessin 1.0.1
dessin is an up to date version of drawing.
It means drawing in French. Note that the binary is
still called "drawing".

This swirl will disappear when Debian updates its package:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1028955

Swirlbuild does not allow package names that already exist
in Debian proper so it is renamed.

WWW: https://maoschanz.github.io/drawing/